fastapi/requirements.txt
#fastapi
#uvicorn[standard]
gunicorn
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
cryptography==42.0.7
python-dotenv==1.0.1
